chore: move plural eslint rules to @grafana/i18n
