+++ packages/grafana-eslint-rules/index.cjs
@@ -7,8 +7,6 @@ const consistentStoryTitles = require('./rules/consistent-story-titles.cjs');
 const noPluginExternalImportPaths = require('./rules/no-plugin-external-import-paths.cjs');
 const noInvalidCssProperties = require('./rules/no-invalid-css-properties.cjs');
 const defineFeatureEvents = require('./rules/define-feature-events.cjs');
-const tPluralDefaults = require('./rules/t-plural-defaults.cjs');
-const transPluralDefaults = require('./rules/trans-plural-defaults.cjs');
 const noRestrictedSyntaxRules = require('./rules/no-restricted-syntax.cjs');
 
 /** @type {import('eslint').Linter.Plugin} */
@@ -23,8 +21,6 @@ module.exports = {
     'no-plugin-external-import-paths': noPluginExternalImportPaths,
     'no-invalid-css-properties': noInvalidCssProperties,
     'define-feature-events': defineFeatureEvents,
-    't-plural-defaults': tPluralDefaults,
-    'trans-plural-defaults': transPluralDefaults,
     ...noRestrictedSyntaxRules.rules,
   },
 };
