+++ packages/grafana-i18n/src/eslint/README.md
@@ -233,3 +233,89 @@ const SomeComponent = () => {
   );
 };
 ```
+
+### `t-plural-defaults`
+
+Enforce the plural-default convention for `t()` calls from `@grafana/i18n`. Whenever the options object contains a `count` key, the call must:
+
+1. Pass an empty string as the positional `defaultValue` (2nd argument).
+2. Include both `defaultValue_one` and `defaultValue_other` in the options object.
+
+`i18next` selects the plural variant from `count` + CLDR rules at runtime, and the `i18next-cli` extractor emits the matching `<key>_<suffix>` entries when both forms are declared in source. Splitting one form into the positional default and the other into the options object makes the intent ambiguous and breaks the extractor's contract that source code is the source of truth for every plural form.
+
+#### Examples
+
+```ts
+// Bad ❌ — positional default carries the "_other" form implicitly
+t('grafana-data.datetime.rangeutils.lastNMinutes', 'Last {{count}} minutes', {
+  count,
+  defaultValue_one: 'Last {{count}} minute',
+});
+
+// Bad ❌ — missing defaultValue_one
+t('foo', '', {
+  count,
+  defaultValue_other: '{{count}} items',
+});
+
+// Good ✅ — every plural form lives in the options object
+t('grafana-data.datetime.rangeutils.lastNMinutes', '', {
+  count,
+  defaultValue_one: 'Last {{count}} minute',
+  defaultValue_other: 'Last {{count}} minutes',
+});
+
+// Good ✅ — no count, so the rule does not apply
+t('common.hello', 'Hello {{name}}', { name });
+```
+
+This rule is intentionally **not auto-fixable**. A complete fix always needs the engineer to supply the singular form (`defaultValue_one`); a partial auto-fix would risk landing half-migrated call sites. Fix violations by hand and verify the singular/plural copy matches the original intent.
+
+### `trans-plural-defaults`
+
+Enforce the plural-default convention for `<Trans>` components from `@grafana/i18n`. Whenever a `<Trans>` element passes `count` — either as a direct prop or as a key inside its `values` prop — it must declare a `tOptions` prop containing both `defaultValue_one` and `defaultValue_other`.
+
+This is the JSX counterpart to [`t-plural-defaults`](#t-plural-defaults). Same reasoning: `i18next` selects the plural variant from `count` + CLDR rules at runtime, and the `i18next-cli` extractor emits `<key>_<suffix>` entries when both plural defaults are declared in source.
+
+#### Examples
+
+```tsx
+// Bad ❌ — count is present, no tOptions at all
+<Trans i18nKey="alerting.alert-instance-extension-point.view-route" count={journeys.length}>
+  View route
+</Trans>
+
+// Bad ❌ — count lives inside values, still no tOptions
+<Trans i18nKey="alerting.triage.showing-groups-count" values={{ count: data.length }}>
+  Showing {'{{count}}'} groups
+</Trans>
+
+// Bad ❌ — tOptions is missing one of the plural forms
+<Trans i18nKey="foo" count={n} tOptions={{ defaultValue_one: 'one' }}>x</Trans>
+
+// Good ✅ — direct count prop + both plural defaults
+<Trans
+  i18nKey="alerting.alert-instance-extension-point.view-route"
+  count={journeys.length}
+  tOptions={{
+    defaultValue_one: 'View route',
+    defaultValue_other: 'View routes',
+  }}
+>
+  View route
+</Trans>
+
+// Good ✅ — count via values + both plural defaults
+<Trans
+  i18nKey="alerting.triage.showing-groups-count"
+  values={{ count: data.length }}
+  tOptions={{
+    defaultValue_one: 'Showing {{count}} group',
+    defaultValue_other: 'Showing {{count}} groups',
+  }}
+>
+  Showing {'{{count}}'} groups
+</Trans>
+```
+
+This rule is intentionally **not auto-fixable** for the same reason as `t-plural-defaults`: the singular and plural copy can't be inferred and must be supplied by the engineer.
